Refactor code and add tests for logic manager
