+++ src/main/java/seedu/ibook/logic/commands/item/UpdateItemCommand.java
@@ -1,4 +1,4 @@
-package seedu.ibook.logic.commands;
+package seedu.ibook.logic.commands.item;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ibook.model.product.Product.PREDICATE_SHOW_ALL_ITEMS;
@@ -9,6 +9,8 @@ import java.util.Optional;
 import seedu.ibook.commons.core.Messages;
 import seedu.ibook.commons.core.index.CompoundIndex;
 import seedu.ibook.commons.util.CollectionUtil;
+import seedu.ibook.logic.commands.Command;
+import seedu.ibook.logic.commands.CommandResult;
 import seedu.ibook.logic.commands.exceptions.CommandException;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.item.ExpiryDate;
