+++ src/main/java/seedu/ibook/logic/commands/product/UpdateAllCommand.java
@@ -1,4 +1,4 @@
-package seedu.ibook.logic.commands;
+package seedu.ibook.logic.commands.product;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ibook.logic.parser.CliSyntax.PREFIX_CATEGORY;
@@ -7,8 +7,10 @@ import static seedu.ibook.logic.parser.CliSyntax.PREFIX_EXPIRY_DATE;
 import java.util.ArrayList;
 import java.util.List;
 
-import seedu.ibook.logic.commands.UpdateCommand.UpdateProductDescriptor;
+import seedu.ibook.logic.commands.Command;
+import seedu.ibook.logic.commands.CommandResult;
 import seedu.ibook.logic.commands.exceptions.CommandException;
+import seedu.ibook.logic.commands.product.UpdateCommand.UpdateProductDescriptor;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.product.Product;
