+++ src/main/java/seedu/ibook/logic/commands/product/UpdateCommand.java
@@ -1,4 +1,4 @@
-package seedu.ibook.logic.commands;
+package seedu.ibook.logic.commands.product;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ibook.logic.parser.CliSyntax.PREFIX_NAME;
@@ -10,6 +10,8 @@ import java.util.Optional;
 import seedu.ibook.commons.core.Messages;
 import seedu.ibook.commons.core.index.Index;
 import seedu.ibook.commons.util.CollectionUtil;
+import seedu.ibook.logic.commands.Command;
+import seedu.ibook.logic.commands.CommandResult;
 import seedu.ibook.logic.commands.exceptions.CommandException;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.item.UniqueItemList;
