+++ src/main/java/seedu/ibook/logic/parser/DeleteItemCommandParser.java
@@ -3,7 +3,7 @@ package seedu.ibook.logic.parser;
 import static seedu.ibook.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.ibook.commons.core.index.CompoundIndex;
-import seedu.ibook.logic.commands.DeleteItemCommand;
+import seedu.ibook.logic.commands.item.DeleteItemCommand;
 import seedu.ibook.logic.parser.exceptions.ParseException;
 
 public class DeleteItemCommandParser implements Parser<DeleteItemCommand> {
