+++ src/main/java/seedu/ibook/logic/parser/IBookParser.java
@@ -5,22 +5,22 @@ import static seedu.ibook.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.ibook.logic.commands.AddCommand;
-import seedu.ibook.logic.commands.AddItemCommand;
 import seedu.ibook.logic.commands.ClearCommand;
 import seedu.ibook.logic.commands.Command;
-import seedu.ibook.logic.commands.DeleteAllCommand;
-import seedu.ibook.logic.commands.DeleteCommand;
-import seedu.ibook.logic.commands.DeleteItemCommand;
 import seedu.ibook.logic.commands.ExitCommand;
 import seedu.ibook.logic.commands.ExpiredCommand;
-import seedu.ibook.logic.commands.FindCommand;
 import seedu.ibook.logic.commands.HelpCommand;
-import seedu.ibook.logic.commands.ListCommand;
 import seedu.ibook.logic.commands.OutOfStockCommand;
-import seedu.ibook.logic.commands.UpdateAllCommand;
-import seedu.ibook.logic.commands.UpdateCommand;
-import seedu.ibook.logic.commands.UpdateItemCommand;
+import seedu.ibook.logic.commands.item.AddItemCommand;
+import seedu.ibook.logic.commands.item.DeleteItemCommand;
+import seedu.ibook.logic.commands.item.UpdateItemCommand;
+import seedu.ibook.logic.commands.product.AddCommand;
+import seedu.ibook.logic.commands.product.DeleteAllCommand;
+import seedu.ibook.logic.commands.product.DeleteCommand;
+import seedu.ibook.logic.commands.product.FindCommand;
+import seedu.ibook.logic.commands.product.ListCommand;
+import seedu.ibook.logic.commands.product.UpdateAllCommand;
+import seedu.ibook.logic.commands.product.UpdateCommand;
 import seedu.ibook.logic.parser.exceptions.ParseException;
 
 /**
