+++ src/main/java/seedu/ibook/logic/parser/UpdateItemCommandParser.java
@@ -5,8 +5,8 @@ import static seedu.ibook.logic.parser.CliSyntax.PREFIX_EXPIRY_DATE;
 import static seedu.ibook.logic.parser.CliSyntax.PREFIX_QUANTITY;
 
 import seedu.ibook.commons.core.index.CompoundIndex;
-import seedu.ibook.logic.commands.UpdateItemCommand;
-import seedu.ibook.logic.commands.UpdateItemCommand.UpdateItemDescriptor;
+import seedu.ibook.logic.commands.item.UpdateItemCommand;
+import seedu.ibook.logic.commands.item.UpdateItemCommand.UpdateItemDescriptor;
 import seedu.ibook.logic.parser.exceptions.ParseException;
 
 public class UpdateItemCommandParser implements Parser<UpdateItemCommand> {
