+++ src/main/java/seedu/ibook/ui/popup/PopupAddProduct.java
@@ -3,7 +3,7 @@ package seedu.ibook.ui.popup;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
-import seedu.ibook.logic.commands.AddCommand;
+import seedu.ibook.logic.commands.product.AddCommand;
 import seedu.ibook.logic.parser.CliSyntax;
 import seedu.ibook.ui.MainWindow;
