+++ src/main/java/seedu/ibook/ui/popup/item/PopupManageItem.java
@@ -4,8 +4,8 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import seedu.ibook.commons.core.index.CompoundIndex;
-import seedu.ibook.logic.commands.DeleteItemCommand;
-import seedu.ibook.logic.commands.UpdateItemCommand;
+import seedu.ibook.logic.commands.item.DeleteItemCommand;
+import seedu.ibook.logic.commands.item.UpdateItemCommand;
 import seedu.ibook.logic.parser.CliSyntax;
 import seedu.ibook.model.item.Item;
 import seedu.ibook.model.product.Product;
