+++ src/test/java/seedu/ibook/logic/LogicManagerTest.java
@@ -1,17 +1,14 @@
 package seedu.ibook.logic;
-/*
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static seedu.ibook.commons.core.Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX;
+import static seedu.ibook.commons.core.Messages.MESSAGE_INVALID_PRODUCT_DISPLAYED_INDEX;
 import static seedu.ibook.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
-
-import static seedu.ibook.logic.commands.CommandTestUtil.ADDRESS_DESC_AMY;
-import static seedu.ibook.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.ibook.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.ibook.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
-
-
+import static seedu.ibook.logic.commands.CommandTestUtil.CATEGORY_FULL_A;
+import static seedu.ibook.logic.commands.CommandTestUtil.DESCRIPTION_FULL_A;
+import static seedu.ibook.logic.commands.CommandTestUtil.NAME_FULL_A;
+import static seedu.ibook.logic.commands.CommandTestUtil.PRICE_FULL_A;
 import static seedu.ibook.testutil.Assert.assertThrows;
-//import static seedu.ibook.testutil.TypicalPersons.AMY;
+import static seedu.ibook.testutil.TypicalProducts.PRODUCT_A;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -20,25 +17,23 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import seedu.ibook.logic.commands.AddCommand;
 import seedu.ibook.logic.commands.CommandResult;
-import seedu.ibook.logic.commands.ListCommand;
 import seedu.ibook.logic.commands.exceptions.CommandException;
+import seedu.ibook.logic.commands.product.AddCommand;
+import seedu.ibook.logic.commands.product.ListCommand;
 import seedu.ibook.logic.parser.exceptions.ParseException;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.ModelManager;
-import seedu.ibook.model.UserPrefs;
 import seedu.ibook.model.ReadOnlyIBook;
-import seedu.ibook.model.person.Person;
-import seedu.ibook.storage.JsonAddressBookStorage;
+import seedu.ibook.model.UserPrefs;
+import seedu.ibook.model.product.Product;
+import seedu.ibook.storage.JsonIBookStorage;
 import seedu.ibook.storage.JsonUserPrefsStorage;
 import seedu.ibook.storage.StorageManager;
-import seedu.ibook.testutil.PersonBuilder;
-
- */
+import seedu.ibook.testutil.ProductBuilder;
 
 public class LogicManagerTest {
-    /*
+
     private static final IOException DUMMY_IO_EXCEPTION = new IOException("dummy exception");
 
     @TempDir
@@ -49,10 +44,10 @@ public class LogicManagerTest {
 
     @BeforeEach
     public void setUp() {
-        JsonAddressBookStorage addressBookStorage =
-                new JsonAddressBookStorage(temporaryFolder.resolve("addressBook.json"));
+        JsonIBookStorage iBookStorage =
+            new JsonIBookStorage(temporaryFolder.resolve("iBook.json"));
         JsonUserPrefsStorage userPrefsStorage = new JsonUserPrefsStorage(temporaryFolder.resolve("userPrefs.json"));
-        StorageManager storage = new StorageManager(addressBookStorage, userPrefsStorage);
+        StorageManager storage = new StorageManager(iBookStorage, userPrefsStorage);
         logic = new LogicManager(model, storage);
     }
 
@@ -65,7 +60,7 @@ public class LogicManagerTest {
     @Test
     public void execute_commandExecutionError_throwsCommandException() {
         String deleteCommand = "delete 9";
-        assertCommandException(deleteCommand, MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
+        assertCommandException(deleteCommand, MESSAGE_INVALID_PRODUCT_DISPLAYED_INDEX);
     }
 
     @Test
@@ -76,27 +71,26 @@ public class LogicManagerTest {
 
     @Test
     public void execute_storageThrowsIoException_throwsCommandException() {
-        // Setup LogicManager with JsonAddressBookIoExceptionThrowingStub
-        JsonAddressBookStorage addressBookStorage =
-                new JsonAddressBookIoExceptionThrowingStub(temporaryFolder.resolve("ioExceptionAddressBook.json"));
+        // Setup LogicManager with JsonIBookIoExceptionThrowingStub
+        JsonIBookStorage iBookStorage =
+            new JsonIBookIoExceptionThrowingStub(temporaryFolder.resolve("ioExceptionIBook.json"));
         JsonUserPrefsStorage userPrefsStorage =
-                new JsonUserPrefsStorage(temporaryFolder.resolve("ioExceptionUserPrefs.json"));
-        StorageManager storage = new StorageManager(addressBookStorage, userPrefsStorage);
+            new JsonUserPrefsStorage(temporaryFolder.resolve("ioExceptionUserPrefs.json"));
+        StorageManager storage = new StorageManager(iBookStorage, userPrefsStorage);
         logic = new LogicManager(model, storage);
 
         // Execute add command
-        String addCommand = AddCommand.COMMAND_WORD + NAME_DESC_AMY + PHONE_DESC_AMY + EMAIL_DESC_AMY
-                + ADDRESS_DESC_AMY;
-        Person expectedPerson = new PersonBuilder(AMY).withTags().build();
+        String addCommand = AddCommand.COMMAND_WORD + NAME_FULL_A + CATEGORY_FULL_A + DESCRIPTION_FULL_A + PRICE_FULL_A;
+        Product expectedProduct = new ProductBuilder(PRODUCT_A).build();
         ModelManager expectedModel = new ModelManager();
-        expectedModel.addPerson(expectedPerson);
+        expectedModel.addProduct(expectedProduct);
         String expectedMessage = LogicManager.FILE_OPS_ERROR_MESSAGE + DUMMY_IO_EXCEPTION;
         assertCommandFailure(addCommand, CommandException.class, expectedMessage, expectedModel);
     }
 
     @Test
-    public void getFilteredPersonList_modifyList_throwsUnsupportedOperationException() {
-        assertThrows(UnsupportedOperationException.class, () -> logic.getFilteredPersonList().remove(0));
+    public void getFilteredProductList_modifyList_throwsUnsupportedOperationException() {
+        assertThrows(UnsupportedOperationException.class, () -> logic.getFilteredIBook().remove(0));
     }
 
     /**
@@ -105,7 +99,7 @@ public class LogicManagerTest {
      * - the feedback message is equal to {@code expectedMessage} <br>
      * - the internal model manager state is the same as that in {@code expectedModel} <br>
      * @see #assertCommandFailure(String, Class, String, Model)
-
+     */
     private void assertCommandSuccess(String inputCommand, String expectedMessage,
             Model expectedModel) throws CommandException, ParseException {
         CommandResult result = logic.execute(inputCommand);
@@ -116,7 +110,7 @@ public class LogicManagerTest {
     /**
      * Executes the command, confirms that a ParseException is thrown and that the result message is correct.
      * @see #assertCommandFailure(String, Class, String, Model)
-
+     */
     private void assertParseException(String inputCommand, String expectedMessage) {
         assertCommandFailure(inputCommand, ParseException.class, expectedMessage);
     }
@@ -124,7 +118,7 @@ public class LogicManagerTest {
     /**
      * Executes the command, confirms that a CommandException is thrown and that the result message is correct.
      * @see #assertCommandFailure(String, Class, String, Model)
-
+     */
     private void assertCommandException(String inputCommand, String expectedMessage) {
         assertCommandFailure(inputCommand, CommandException.class, expectedMessage);
     }
@@ -132,10 +126,10 @@ public class LogicManagerTest {
     /**
      * Executes the command, confirms that the exception is thrown and that the result message is correct.
      * @see #assertCommandFailure(String, Class, String, Model)
-
+     */
     private void assertCommandFailure(String inputCommand, Class<? extends Throwable> expectedException,
             String expectedMessage) {
-        Model expectedModel = new ModelManager(model.getAddressBook(), new OldUserPrefs());
+        Model expectedModel = new ModelManager(model.getIBook(), new UserPrefs());
         assertCommandFailure(inputCommand, expectedException, expectedMessage, expectedModel);
     }
 
@@ -145,7 +139,7 @@ public class LogicManagerTest {
      * - the resulting error message is equal to {@code expectedMessage} <br>
      * - the internal model manager state is the same as that in {@code expectedModel} <br>
      * @see #assertCommandSuccess(String, String, Model)
-
+     */
     private void assertCommandFailure(String inputCommand, Class<? extends Throwable> expectedException,
             String expectedMessage, Model expectedModel) {
         assertThrows(expectedException, expectedMessage, () -> logic.execute(inputCommand));
@@ -154,17 +148,15 @@ public class LogicManagerTest {
 
     /**
      * A stub class to throw an {@code IOException} when the save method is called.
-
-    private static class JsonAddressBookIoExceptionThrowingStub extends JsonAddressBookStorage {
-        private JsonAddressBookIoExceptionThrowingStub(Path filePath) {
+     */
+    private static class JsonIBookIoExceptionThrowingStub extends JsonIBookStorage {
+        private JsonIBookIoExceptionThrowingStub(Path filePath) {
             super(filePath);
         }
 
         @Override
-        public void saveAddressBook(ReadOnlyAddressBook addressBook, Path filePath) throws IOException {
+        public void saveIBook(ReadOnlyIBook iBook, Path filePath) throws IOException {
             throw DUMMY_IO_EXCEPTION;
         }
     }
-
-     */
 }
