+++ src/test/java/seedu/ibook/logic/commands/AddCommandIntegrationTest.java
@@ -7,6 +7,7 @@ import static seedu.ibook.testutil.TypicalProducts.getTypicalIBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.ibook.logic.commands.product.AddCommand;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.ModelManager;
 import seedu.ibook.model.UserPrefs;
