+++ src/test/java/seedu/ibook/logic/commands/DeleteAllCommandTest.java
@@ -8,6 +8,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.ibook.logic.commands.product.DeleteAllCommand;
 import seedu.ibook.model.IBook;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.ModelManager;
