+++ src/test/java/seedu/ibook/logic/commands/UpdateAllCommandTest.java
@@ -16,7 +16,9 @@ import static seedu.ibook.testutil.TypicalProducts.getTypicalIBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.ibook.logic.commands.UpdateCommand.UpdateProductDescriptor;
+import seedu.ibook.logic.commands.product.UpdateAllCommand;
+import seedu.ibook.logic.commands.product.UpdateCommand;
+import seedu.ibook.logic.commands.product.UpdateCommand.UpdateProductDescriptor;
 import seedu.ibook.model.IBook;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.ModelManager;
