+++ src/test/java/seedu/ibook/logic/parser/IBookParserTest.java
@@ -16,18 +16,18 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.ibook.logic.commands.AddCommand;
 import seedu.ibook.logic.commands.ClearCommand;
-import seedu.ibook.logic.commands.DeleteAllCommand;
-import seedu.ibook.logic.commands.DeleteCommand;
 import seedu.ibook.logic.commands.ExitCommand;
 import seedu.ibook.logic.commands.ExpiredCommand;
-import seedu.ibook.logic.commands.FindCommand;
-import seedu.ibook.logic.commands.ListCommand;
 import seedu.ibook.logic.commands.OutOfStockCommand;
-import seedu.ibook.logic.commands.UpdateAllCommand;
-import seedu.ibook.logic.commands.UpdateCommand;
-import seedu.ibook.logic.commands.UpdateCommand.UpdateProductDescriptor;
+import seedu.ibook.logic.commands.product.AddCommand;
+import seedu.ibook.logic.commands.product.DeleteAllCommand;
+import seedu.ibook.logic.commands.product.DeleteCommand;
+import seedu.ibook.logic.commands.product.FindCommand;
+import seedu.ibook.logic.commands.product.ListCommand;
+import seedu.ibook.logic.commands.product.UpdateAllCommand;
+import seedu.ibook.logic.commands.product.UpdateCommand;
+import seedu.ibook.logic.commands.product.UpdateCommand.UpdateProductDescriptor;
 import seedu.ibook.logic.parser.exceptions.ParseException;
 import seedu.ibook.model.product.Product;
 import seedu.ibook.model.product.filters.AttributeFilter;
